Fixed checksum generation and parsing bug, adding unit test. Extended positions to 32 bit integers from 16 bit integers

+++ microcontroller/libs/rotary_controller.hpp
@@ -6,19 +6,19 @@
 
 class RotaryController {
   private:
-    int16_t target_steps;
+    int32_t target_steps;
   public:
 
     RotaryController():
       target_steps(0){
     }
 
-    void set_target_steps(int16_t target_steps_){
+    void set_target_steps(int32_t target_steps_){
       target_steps = target_steps_;
     }
 
     void control(RotaryStepper& rotary_stepper) {
-      // error is 32 bit as difference of 16 bits can be bigger than 16 bits
+      // error is 32 bit as difference of 32 bit values that are actually very close to 16 so they won't overflow
       int32_t error = rotary_stepper.get_steps() - target_steps;
       if(error < 0) {
         // move anticlockwise
